pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GTFS-flex-to-GOFS"
version = "0.5.0"
description = "Convert GTFS Flex to GOFS"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nicholas Paun"},
    {name = "Jonathan Milot"}
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.8"
dependencies = [
    "py-gtfs-loader @ git+https://github.com/transitapp/py-gtfs-loader@v0.1.15",
    "shapely>=2.0.4"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "flake8"
]

[project.urls]
Homepage = "https://github.com/TransitApp/GTFS-flex-to-GOFS"
Repository = "https://github.com/TransitApp/GTFS-flex-to-GOFS"

[tool.setuptools.packages.find]
include = ["gtfs_flex_to_gofs*"]

[project.scripts]
gtfs-flex-to-gofs = "gtfs_flex_to_gofs.__main__:main"